8_Building_Your_First_GenAI_Application_with_AWS_Data_Foundations/8.3_Building_Your_Life_Insurance_Agent/app/streamlit/requirements.txt
# app/streamlit/requirements.txt

# Core dependencies
streamlit
boto3
python-dotenv

# AWS/Authentication
streamlit-cognito-auth
botocore

# Document processing
langchain-community
amazon-textract-caller

# Bedrock backoff and retry
tenacity
backoff